заготовки к тестам
