+++ main.py
@@ -1,7 +1,4 @@
 from src.interaction import user_interaction
+
 if __name__ == "__main__":
-
-
     user_interaction()
-
-
